Create board and shape pieces
The board generates based off a random generator. The pieces can be dragged
using the mouse and dropped on the tiles. The tile can do a comparison on the
shape dropped against its own shape. It just console logs if it matches or not.
At this point, I can allow the shape to stay or be sent back to the tray.